requirements.txt
fastapi
uvicorn 
firebase-admin
pydantic
pydantic[email]
python-dotenv
